requirements.txt
streamlit==1.28.0
openai==1.3.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
python-docx==1.0.1
PyPDF2==3.0.1
nltk==3.8.1
spacy==3.6.1
transformers==4.35.0
sentence-transformers==2.2.2
langchain==0.0.335
python-dotenv==1.0.0
plotly==5.17.0